Refactor existing barcode scan endpoint
- Break out into new file just for purchase orders
