+++ lib/api.dart
@@ -330,7 +330,11 @@ class InvenTreeAPI {
   // Does the server support extra fields on stock adjustment actions?
   bool get supportsStockAdjustExtraFields => isConnected() && apiVersion >= 133;
 
+  // Does the server support receiving items against a PO using barcodes?
   bool get supportsBarcodePOReceiveEndpoint => isConnected() && apiVersion >= 139;
+
+  // Does the server support adding line items to a PO using barcodes?
+  bool get supportsBarcodePOAddLineEndpoint => isConnected() && apiVersion >= 153;
 
   // Cached list of plugins (refreshed when we connect to the server)
   List<InvenTreePlugin> _plugins = [];
