+++ lib/barcode/barcode.dart
@@ -7,7 +7,6 @@ import "package:one_context/one_context.dart";
 
 
 import "package:inventree/api.dart";
-import "package:inventree/api_form.dart";
 import "package:inventree/helpers.dart";
 import "package:inventree/l10.dart";
 
@@ -219,6 +218,7 @@ class BarcodeScanHandler extends BarcodeHandler {
         case "purchaseorder":
           await handlePurchaseOrder(pk);
           return;
+          // TODO: Handle manufacturer part
         default:
           // Fall through to failure state
           break;
@@ -474,116 +474,6 @@ class ScanParentLocationHandler extends BarcodeScanStockLocationHandler {
         );
         return false;
     }
-  }
-}
-
-
-/*
- * Barcode handler class for scanning a supplier barcode to receive a part
- *
- * - The class can be initialized by optionally passing a valid, placed PurchaseOrder object
- * - Expects to scan supplier barcode, possibly containing order_number and quantity
- * - If location or quantity information wasn't provided, show a form to fill it in
- */
-class POReceiveBarcodeHandler extends BarcodeHandler {
-
-  POReceiveBarcodeHandler({this.purchaseOrder, this.location});
-
-  InvenTreePurchaseOrder? purchaseOrder;
-  InvenTreeStockLocation? location;
-
-  @override
-  String getOverlayText(BuildContext context) => L10().barcodeReceivePart;
-
-  @override
-  Future<void> processBarcode(String barcode,
-      {String url = "barcode/po-receive/",
-      Map<String, dynamic> extra_data = const {}}) {
-
-    final po_extra_data = {
-      "purchase_order": purchaseOrder?.pk,
-      "location": location?.pk,
-      ...extra_data,
-    };
-
-    return super.processBarcode(barcode, url: url, extra_data: po_extra_data);
-  }
-
-  @override
-  Future<void> onBarcodeMatched(Map<String, dynamic> data) async {
-      if (!data.containsKey("lineitem")) {
-        return onBarcodeUnknown(data);
-      }
-
-      barcodeSuccessTone();
-      showSnackIcon(L10().receivedItem, success: true);
-  }
-
-  @override
-  Future<void> onBarcodeUnhandled(Map<String, dynamic> data) async {
-    if (!data.containsKey("action_required") || !data.containsKey("lineitem")) {
-      return super.onBarcodeUnhandled(data);
-    }
-
-    final lineItemData = data["lineitem"] as Map<String, dynamic>;
-    if (!lineItemData.containsKey("pk") || !lineItemData.containsKey("purchase_order")) {
-      barcodeFailureTone();
-      showSnackIcon(L10().missingData, success: false);
-    }
-
-    // Construct fields to receive
-    Map<String, dynamic> fields = {
-      "line_item": {
-        "parent": "items",
-        "nested": true,
-        "hidden": true,
-        "value": lineItemData["pk"] as int,
-      },
-      "quantity": {
-        "parent": "items",
-        "nested": true,
-        "value": lineItemData["quantity"] as double?,
-      },
-      "status": {
-        "parent": "items",
-        "nested": true,
-      },
-      "location": {
-        "value": lineItemData["location"] as int?,
-      },
-      "barcode": {
-        "parent": "items",
-        "nested": true,
-        "hidden": true,
-        "type": "barcode",
-        "value": data["barcode_data"] as String,
-      }
-    };
-
-    final context = OneContext().context!;
-    final purchase_order_pk = lineItemData["purchase_order"];
-    final receive_url = "${InvenTreePurchaseOrder().URL}${purchase_order_pk}/receive/";
-
-    launchApiForm(
-      context,
-      L10().receiveItem,
-      receive_url,
-      fields,
-      method: "POST",
-      icon: FontAwesomeIcons.rightToBracket,
-      onSuccess: (data) async {
-        showSnackIcon(L10().receivedItem, success: true);
-      }
-    );
-  }
-
-  @override
-  Future<void> onBarcodeUnknown(Map<String, dynamic> data) async {
-    barcodeFailureTone();
-    showSnackIcon(
-      data["error"] as String? ?? L10().barcodeError,
-      success: false
-    );
   }
 }
